pointerevents/pointerevent_gotpointercapture_before_first_pointerevent.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: gotpiontercapture is fired first.</title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <meta name="assert" content="After setting capture, the gotpointercapture dispatched to the capturing element before any other event is fired."/>

        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
        <script type="text/javascript">
            setup({ explicit_done: true });

            var detected_pointertypes = {};
            add_completion_callback(showPointerTypes);

            var target0;
            var listener;
            var pointerover_event;
            var detected_pointerEvents = new Array();
            var eventRcvd = false;
            var test_pointerCapture;

            function run() {

                target0 = document.getElementById("target0");
                target0.style["touchAction"] = "none";
                listener = document.getElementById("listener");

                test_pointerCapture = async_test("gotpointercapture event");  // set up test harness

                // listen to all events
                for (var i = 0; i < All_Pointer_Events.length; i++) {

                    on_event(listener, All_Pointer_Events[i], function (event) {
                        if (event.type == "gotpointercapture") {
                            // if any events have been received with same pointerId before gotpointercapture, then fail
                            var eventsRcvd_str = "";
                            if (!eventRcvd) {
                                eventsRcvd_str = "Events received before gotpointercapture: ";
                                for (var i = 0; i < detected_pointerEvents.length; i++) {
                                    eventsRcvd_str += detected_pointerEvents[i] +", ";
                                }
                            }
                            test_pointerCapture.step(function () {
                                assert_false(eventRcvd, "no other events should be received before gotpointercapture." + eventsRcvd_str);
                                assert_equals(event.pointerId, pointerover_event.pointerId, "pointerID is same for pointerover and gotpointercapture");
                            });

                            detected_pointertypes[event.pointerType] = true;
                            test_pointerCapture.done(); // complete test
                        }
                        else {
                            if (pointerover_event.pointerId === event.pointerId) {
                                detected_pointerEvents.push(event.type);
                                eventRcvd = true;
                            }
                        }
                    });
                }

                // set pointer capture
                on_event(target0, "pointerover", function (event) {
                    pointerover_event = event;
                    listener.setPointerCapture(event.pointerId);
                });
            }

            function showPointerTypes() {
                var complete_notice = document.getElementById("complete-notice");
                var pointertype_log = document.getElementById("pointertype-log");
                var pointertypes = Object.keys(detected_pointertypes);
                pointertype_log.innerHTML = pointertypes.length ? pointertypes.join(",") : "(none)";
                if (test_pointerCapture.status == test_pointerCapture.NOTRUN
                    || test_pointerCapture.status == test_pointerCapture.TIMEOUT) {
                    complete_notice.innerHTML = "Test FAIL due to time out.  This is due to the browser not supporting pointer, or the manual steps were not run to completion.";
                    complete_notice.style.background = "red";
                }
                complete_notice.style.display = "block";
            }
        </script>

        <style>
            div {
                margin: 0em;
                padding: 2em;
            }

            #target0 {
                background: yellow;
                border: 1px solid orange;
            }
            #listener {
                background: orange;
                border: 1px solid orange;
                position: absolute;
                top: -100px;
            }

            #complete-notice {
                border: 1px solid #0a0;
                display: none;
            }

            #pointertype-log {
                font-weight: bold;
            }
        </style>
    </head>
    <body onload="run()">
        <h1> Pointer Event: Dispatch gotpointercapture event</h1>
        <h4> 
            Test Description: 
            After pointer capture is set for a pointer, and prior to dispatching the first event for the pointer, the gotpointercapture 
            event must be dispatched to the element that is receiving the pointer capture. 
        </h4>
        <br />
        <div id="target0">
            Use the mouse, touch or pen to move over this box.
        </div>
        <div id="listener">Do not hover over or touch this element. </div>
        <div id="complete-notice">
            <p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
            <p>Would you like to <a href="#">run the tests again</a> with a different pointer type?</p>
        </div>

        <div id="log"></div>
    </body>
</html>
